Add a method for a corporate id, use ACN and fall back to ABN
This is so I show a number for the company under their name, and
then also easily use that number for their id/url.

It also shows 'no_corporate_id' if they don't have one, this would
be very unusual and would warrant investigation. Why are entities
without an abn or acn being listed?

I thought about raising an error in that situation, but that seems
like a nasty experience for people.